requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
face-recognition==1.3.0
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.4
python-dotenv==1.0.0
email-validator==2.0.0
gunicorn==21.2.0
python-multipart==0.0.6
twilio==8.12.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
celery==5.3.4
redis==5.0.1
APScheduler==3.10.4
